Refactor contractstate
