+++ neo3/storage/__init__.py
@@ -14,5 +14,4 @@ from .cache import (Trackable,
 from .snapshot import CloneSnapshot, Snapshot
 from .storageitem import StorageItem, StorageFlags
 from .storagekey import StorageKey
-from .contractstate import ContractState
 from .utils import NEOByteCompare, NEOSeekSort, create_find_prefix
