+++ tests/contracts/native/test_policy.py
@@ -31,7 +31,7 @@ def test_native_contract(contract_hash: types.UInt160, operation: str, args=None
     # storing the current script in a contract otherwise "System.Contract.Call" will fail its checks
     nef = contracts.NEF(script=sb.to_array())
     manifest = contracts.ContractManifest("test_contract")
-    contract = storage.ContractState(engine.snapshot.contract_id + 1, nef, manifest, 0, to_script_hash(nef.script))
+    contract = contracts.ContractState(engine.snapshot.contract_id + 1, nef, manifest, 0, to_script_hash(nef.script))
     engine.snapshot.contracts.put(contract)
 
     return engine
@@ -124,7 +124,7 @@ class TestPolicyContract(unittest.TestCase):
         nef = contracts.NEF(script=sb.to_array())
         manifest = contracts.ContractManifest("test_contract")
         sender = engine.script_container.script_hashes[0]
-        contract = storage.ContractState(1, nef, manifest, 0, contract_hash(sender, nef.checksum, manifest.name))
+        contract = contracts.ContractState(1, nef, manifest, 0, contract_hash(sender, nef.checksum, manifest.name))
         engine.snapshot.contracts.put(contract)
 
         engine.execute()
