+++ tests/contracts/test_contractstate.py
@@ -1,5 +1,5 @@
 import unittest
-from neo3 import storage, contracts
+from neo3 import contracts
 from neo3.core import types
 
 
@@ -7,7 +7,7 @@ class ContractStateTestCase(unittest.TestCase):
     def test_contractstate_clone(self):
         manifest = contracts.ContractManifest()
         nef = contracts.NEF()
-        state = storage.ContractState(1, nef, manifest, 0, types.UInt160.zero())
+        state = contracts.ContractState(1, nef, manifest, 0, types.UInt160.zero())
         clone = state.clone()
         self.assertNotEqual(id(state), id(clone))
         self.assertNotEqual(id(state.manifest), id(clone.manifest))
@@ -15,12 +15,12 @@ class ContractStateTestCase(unittest.TestCase):
     def test_equals(self):
         manifest = contracts.ContractManifest()
         nef = contracts.NEF()
-        state = storage.ContractState(1, nef, manifest, 0, types.UInt160.zero())
+        state = contracts.ContractState(1, nef, manifest, 0, types.UInt160.zero())
         clone = state.clone()
         self.assertEqual(state, clone)
 
         nef2 = contracts.NEF()
-        state2 = storage.ContractState(2, nef2, manifest, 0, types.UInt160(b'\x01' * 20))
+        state2 = contracts.ContractState(2, nef2, manifest, 0, types.UInt160(b'\x01' * 20))
         self.assertNotEqual(state, state2)
         self.assertNotEqual(state, None)
         self.assertNotEqual(state, object())
